github-code-review-poc/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic-settings==2.3.4
PyGithub==2.3.0
apscheduler==3.10.4
anthropic==0.28.0
cryptography==42.0.8
python-dotenv==1.0.1
httpx==0.27.0
